refactor and add news.service
