+++ oser-app/src/app/showcase-site/news-page/news-item/news-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { News } from '../../news';
+import { News } from '../../shared/news.model';
 
 @Component({
   selector: 'app-news-item',
