+++ oser-app/src/app/showcase-site/shared/mock-news.ts
@@ -1,4 +1,4 @@
-import { News } from './news';
+import { News } from './news.model';
 
 /*
 This is a mock array of News objects.
