+++ oser-app/src/app/showcase-site/shared/news.model.ts
