+++ oser-app/src/app/showcase-site/shared/news.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { NewsService } from './news.service';
+
+describe('NewsService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [NewsService]
+    });
+  });
+
+  it('should ...', inject([NewsService], (service: NewsService) => {
+    expect(service).toBeTruthy();
+  }));
+});
