+++ oser-app/src/app/showcase-site/showcase-site.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
-import jump from 'jump.js';
 
 @Component({
   selector: 'app-showcase-site',
@@ -16,7 +15,8 @@ export class ShowcaseSiteComponent implements OnInit {
       if (!(event instanceof NavigationEnd)) {
         return;
       }
-      jump('#top', {duration: 500});
+      // Scroll to top of window
+      window.scrollTo(0, 0);
     });
   }
