+++ oser-app/src/app/showcase-site/showcase-site.module.ts
@@ -1,17 +1,25 @@
 import { NgModule } from '@angular/core';
+
+// Modules
 import { CommonModule } from '@angular/common';
+import { ShowcaseRoutingModule } from './showcase-routing.module';
+
+// Components
 import { NavbarComponent } from './navbar/navbar.component';
 import { ShowcaseSiteComponent } from './showcase-site.component';
 import { LandingPageComponent } from './landing-page/landing-page.component';
 import { NewsCardComponent } from './landing-page/news-card/news-card.component';
 import { FooterComponent } from './footer/footer.component';
-import { ShowcaseRoutingModule } from './showcase-routing.module';
 import { DonatePageComponent } from './donate-page/donate-page.component';
 import { ContactPageComponent } from './contact-page/contact-page.component';
 import { ActionsPageComponent } from './actions-page/actions-page.component';
 import { NewsPageComponent } from './news-page/news-page.component';
 import { AboutPageComponent } from './about-page/about-page.component';
 import { NewsItemComponent } from './news-page/news-item/news-item.component';
+
+// Services
+import { NewsService } from './shared/news.service';
+
 
 @NgModule({
   imports: [
@@ -31,6 +39,7 @@ import { NewsItemComponent } from './news-page/news-item/news-item.component';
     AboutPageComponent,
     NewsItemComponent
   ],
+  providers: [NewsService],
   exports: [
     ShowcaseSiteComponent
   ]
